continuing to try to work on 1st map
Still can’t get this to display or load. I tried to have the map load
everything in a specific order as demonstrated in last class, but that
hasn’t worked.
I want to at least get this to display the content I am trying to work
with so that I can experiment.